setup: install Windows DLLs in correct location
The packages for Windows include pre-built DLLs for OpenSSL that
should be installed in the sslpsk Python package.  Previously, we used
the `data_files` setuptools option to install these. However, it only
worked for the eggs, not wheels. In the wheels, the OpenSSL DLLs were
installed in the wrong location (python27/sslpsk/ instead of
python27/Lib/site-packages/sslpsk).

This commit uses the `package_data` setuptools option, instead of
`data_files`, to ensure that the files are installed inside the
package.  To support this, the files must must be copied into the
package source tree during the build.