unittest: added more test cases
Added tests for different ciphers and different TLS Protocol versions.
Unfortenatlty, the TLS Protocol setting "ssl.PROTOCOL_TLS",
which is the Python ssl module default
and with which the ssl modul should choose the best matching protocol version
fails for the sslpsk client side.
Therefore we mark the test using this as "expected failure".

To improve readablity of the test results,
Python warnings will not be displayed (warnings="ignored").
Also test should best be started in buffered mode ("-b" on command line).

Code is formated by the tool "black".